# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/bom_byproducts:
    get:
      tags:
        - BomByproducts
      summary: List BomByproducts
      description: Get a paginated list of BomByproducts with optional filtering and sorting
      operationId: listBomByproducts
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BomByproductList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - BomByproducts
      summary: Create BomByproduct
      description: Create a new BomByproduct
      operationId: createBomByproduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBomByproductRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BomByproduct'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/bom_subcontractors:
    get:
      tags:
        - BomSubcontractors
      summary: List BomSubcontractors
      description: Get a paginated list of BomSubcontractors with optional filtering and sorting
      operationId: listBomSubcontractors
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BomSubcontractorList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - BomSubcontractors
      summary: Create BomSubcontractor
      description: Create a new BomSubcontractor
      operationId: createBomSubcontractor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBomSubcontractorRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BomSubcontractor'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    BomByproduct:
      type: object
      required:
        - id
        - bom_id
        - item_id
        - quantity
        - cost_share
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          nullable: true
          type: string
          format: uuid
        bom_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        product_category_id:
          nullable: true
          type: string
          format: uuid
        quantity:
          type: string
        cost_share:
          type: string
        metadata:
          type: object
...
    CreateBomByproductRequest:
      type: object
      required:
        - bom_id
        - item_id
        - quantity
        - cost_share
        - metadata
      properties:
        company_id:
          nullable: true
          type: string
          format: uuid
        bom_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        product_category_id:
          nullable: true
          type: string
          format: uuid
        quantity:
          type: string
        cost_share:
          type: string
        metadata:
          type: object
...
    BomByproductList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BomByproduct'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    BomSubcontractor:
      type: object
      required:
        - id
        - bom_id
        - partner_id
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          nullable: true
          type: string
          format: uuid
        bom_id:
          type: string
          format: uuid
        partner_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateBomSubcontractorRequest:
      type: object
      required:
        - bom_id
        - partner_id
        - metadata
      properties:
        company_id:
          nullable: true
          type: string
          format: uuid
        bom_id:
          type: string
          format: uuid
        partner_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    BomSubcontractorList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BomSubcontractor'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []